+++ docs/Changelog.md
@@ -17,6 +17,7 @@
   - `Parameters.POLY_SYMBOL = true;`
 - Map fractional zoom [#487](https://github.com/mapsforge/vtm/issues/487)
 - Render theme fallback internal resources [#477](https://github.com/mapsforge/vtm/issues/477)
+- Fix ItemizedLayer synchronization [#507](https://github.com/mapsforge/vtm/issues/507)
 - Fix FadeStep alpha interpolation [#486](https://github.com/mapsforge/vtm/issues/486)
 - Fix libGDX flickering [#148](https://github.com/mapsforge/vtm/issues/148) [#149](https://github.com/mapsforge/vtm/issues/149)
 - JTS (LocationTech) [#484](https://github.com/mapsforge/vtm/issues/484)